[CausalLM] fix type to size_t
- This patch aims to resolve 'Multiplication result converted to larger
  type' issue.
- This patch updates the data type, defined in unsigned int to size_t.

+++ Applications/CausalLM/layers/tie_word_embedding.cpp
@@ -215,9 +215,9 @@ void TieWordEmbedding::incremental_forwarding_embedding(
     throw std::invalid_argument(
       "Tieword embedding is not supported yet for the data type");
 
-  unsigned int b_size = input_.batch();
+  size_t b_size = input_.batch();
 
-  for (unsigned int b = 0; b < b_size; ++b) {
+  for (size_t b = 0; b < b_size; ++b) {
     float *in_data =
       input_.getAddress<float>(b * input_.getDim().getFeatureLen());
